local pywebview

+++ helpers/pythonPublish.js
@@ -10,4 +10,5 @@ for (let line in pySetup) {
   }
 }
 fs.writeFileSync("./module/setup.py", pySetup.join("\r\n"))
-fs.rmdirSync("./module/dist/", { recursive: true })
+if (fs.existsSync("./module/dist/"))
+  fs.rmdirSync("./module/dist/", { recursive: true })
